Convert Main to FXML (janky)
