+++ client/src/main/resources/EventWidget.fxml
@@ -8,8 +8,8 @@
 <fx:root type="javafx.scene.layout.VBox" xmlns:fx="http://javafx.com/fxml">
     <HBox>
         <VBox>
-            <Label fx:id="titleLabel"/>
-            <Label fx:id="descLabel"/>
+            <Label fx:id="titleLabel" minWidth="200"/>
+            <Label fx:id="descLabel" minWidth="200"/>
         </VBox>
         <Button fx:id="doneButton">
             <graphic>
