+++ client/src/main/resources/EventsView.fxml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.layout.*?>
+
+<BorderPane fx:id="bPane" xmlns="http://javafx.com/javafx/17.0.12" xmlns:fx="http://javafx.com/fxml/1" fx:controller="oop.tegevusteplaneerija.client.EventViewsController">
+    <VBox fx:id="events">
+    </VBox>
+</BorderPane>
